Feature/continue react front end

+++ api/routes/site.js
@@ -786,38 +786,52 @@ async function Site(router, sequelizeObjects) {
    * Create license plate
    */
   router.post('/manage/licence/plates', async (req, res) => {
-    sequelizeObjects.Plate.create({
-      licence_plate: req.body.licence_plate, owner_name: req.body.owner_name, enabled: 1
-    }).then(result => {
-      res.status(200);
-      res.send('New licence plate added.');
+    const licensePlate = req.body.licence_plate;
+    sequelizeObjects.Plate.findAll({
+      attributes: ['id',],
+      where: {licence_plate: licensePlate},
+    }).then(rows => {
+      if (rows.length > 0) {
+        res.status(409);
+        res.send('Plate already exists in records');
+      } else {
+        sequelizeObjects.Plate.create({
+          licence_plate: req.body.licence_plate, owner_name: req.body.owner_name, enabled: 1
+        }).then(result => {
+          res.status(200);
+          res.send('New licence plate added.');
+        }).catch(error => {
+          res.status(500);
+          res.send('Error adding licence plate. ' + error);
+        });
+      }
     }).catch(error => {
       res.status(500);
-      res.send('Error adding licence plate. ' + error);
+      res.send(error);
     });
   });
 
   /**
    * Delete license plate
    */
-  router.delete('/manage/licence/plates', async (req, res) => {
-    console.log('Removing ', req.body.data.id);
-    /*
-    sequelizeObjects.Plate.destroy({
-      where: {id: req.body.id,},
-    }).then(result => {
-      if (result === 1) {
-        res.status(200);
-        res.send('Removed licence plate.');
-      } else {
-        res.status(500);
-        res.send('Error removing licence plate. ');
-      }
-    });
-     */
-
-    res.status(200);
-    res.send('Removed licence plate.');
+  router.delete('/manage/licence/plates/:id', async (req, res) => {
+    const id = Number(req.params.id);
+    if (id > 0) {
+      sequelizeObjects.Plate.destroy({
+        where: {id: id},
+      }).then(result => {
+        if (result === 1) {
+          res.status(200);
+          res.send('Removed licence plate.');
+        } else {
+          res.status(500);
+          res.send('Error removing licence plate. ');
+        }
+      });
+    } else {
+      res.status(500);
+      res.send('No proper :id parameter provided for delete.');
+    }
   });
 
   /**
